requirements.txt
weaviate-client>=4.9.0
python-telegram-bot>=21.0
google-genai>=1.0.0
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
